backend/pyproject.toml
[project]
name = "test_case_ccr"
version = "1.0.0"
description = "Тестовое задание для ЦЦР"
authors = [{name = "Evgeniy Shaleynikov", email = "shaleinikov@yandex.ru"}]
readme = "README.md"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "3.12.*"
celery = "^5.4.0"
certifi = "^2023.5.7"
Django = "~5.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
django-filter = "^25.2"
django-imagekit = "^4.1.0"
django-redis = "^5.4.0"
drf-spectacular = "^0.29.0"
factory-boy = "^3.2.1"
gunicorn = "^23.0.0"
pyyaml = "^6.0"
pika = "^1.2.0"
psycopg2-binary = "^2.9.6"
pytest = "^9.0.3"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"
pytest-mock = "^3.6.1"
pytest-sugar = "^0.9.4"
redis = "^5.0.0"
urllib3 = "^2.6.3"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.1.0"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
